requirements.txt
Flask==2.0.1
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc
Werkzeug==2.2.2
numpy==1.21.2
opencv-python-headless==4.5.3.56
Pillow==8.2.0
simple-lama-inpainting
